LPS-96525 Extract inner class to buy space.
